+++ src/lib/services/bank/rest.ts
@@ -16,7 +16,7 @@ export const getBalancesRest = async (
       .get(`${endpoint}/cosmos/bank/v1beta1/balances/${encodeURI(address)}`, {
         params: {
           "pagination.key": paginationKey,
-          "pagination.limit": "100",
+          "pagination.limit": "25",
         },
       })
       .then(({ data }) => parseWithError(zBalancesReponseRest, data));