+++ CHANGELOG.md
@@ -41,6 +41,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+### Bug fixes
+
+## v1.17.0
+
+### Features
+
+### Improvements
+
 - [#1511](https://github.com/alleslabs/celatone-frontend/pull/1511) Upgrade dependencies to fix security vulnerabilities (next, react, axios, sentry, typescript)
 
 ### Bug fixes